+++ test/python/config.py
@@ -2,9 +2,6 @@ from zensols.config import ExtendedInterpolationConfig
 
 
 class AppConfig(ExtendedInterpolationConfig):
-    def __init__(self, *args, **kwargs):
-        super(AppConfig, self).__init__(*args, default_expect=True, **kwargs)
-
     @staticmethod
     def instance():
         return AppConfig('test-resources/db.conf')